INTTEMPLATES-12 - Fix Adapter Template Gradle Tasks
* Fix Gradle tasks: api, reference, dist
* Improve Maven meta data generation in **publish-maven.gradle**
* Add instructions to customize Maven Meta data using Gradle task *install*
* Allow unlimited package nesting
* Improve README.md file
