+++ si-template-projects/adapter/build.gradle
@@ -18,52 +18,21 @@ group = 'SI-TEMPLATE-GROUP-ID'
 
 repositories {
 	maven { url 'http://repo.springsource.org/libs-milestone' }
-	maven { url 'http://repo.springsource.org/plugins-release' } // for bundlor
+	maven { url 'http://repo.springsource.org/plugins-release' }
 }
 
-// ensure JDK 5 compatibility (GRADLE-18; INT-1578)
-sourceCompatibility=1.5
-targetCompatibility=1.5
+sourceCompatibility=1.6
+targetCompatibility=1.6
 
 ext {
-	aspectjVersion = '1.6.8'
-	cglibVersion = '2.2'
-	commonsNetVersion = '3.0.1'
-	easymockVersion = '2.3'
-	groovyVersion = '1.8.5'
-	jacksonVersion = '1.9.2'
-	javaxActivationVersion = '1.1.1'
-	junitVersion = '4.8.2'
+	junitVersion = '4.10'
 	log4jVersion = '1.2.12'
 	mockitoVersion = '1.9.0'
-	// When changing Spring Versions - don't forget to update bundlor ranges
-	springVersion = '3.1.1.RELEASE'
+	springVersion = '3.1.2.RELEASE'
 	springIntegrationVersion = '2.1.2.RELEASE'
-	springAmqpVersion = '1.1.0.RELEASE'
-	springDataMongoVersion = '1.1.0.M1'
-	springDataRedisVersion = '1.0.0.RELEASE'
-	springGemfireVersion = '1.1.1.RELEASE'
-	springSecurityVersion = '3.1.0.RELEASE'
-	springSocialTwitterVersion = '1.0.1.RELEASE'
-	springWsVersion = '2.0.3.RELEASE'
+	
+	idPrefix = 'siAdapterLowerPrefix'
 }
-
-dependencies {
-	compile "org.springframework.integration:spring-integration-core:$springIntegrationVersion"
-	compile "org.springframework:spring-tx:$springVersion"
-	compile("org.codehaus.jackson:jackson-mapper-asl:$jacksonVersion", optional)
-	compile("org.springframework.amqp:spring-rabbit:$springAmqpVersion") {
-		exclude group: 'org.springframework', module: 'spring-aop'
-		exclude group: 'org.springframework', module: 'spring-beans'
-		exclude group: 'org.springframework', module: 'spring-context'
-		exclude group: 'org.springframework', module: 'spring-core'
-		exclude group: 'org.springframework', module: 'spring-oxm'
-		exclude group: 'org.springframework', module: 'spring-tx'
-	}
-
-	testCompile "org.springframework.integration:spring-integration-test:$springIntegrationVersion"
-}
-
 
 eclipse {
 	project {
@@ -85,18 +54,16 @@ configurations {
 	jacoco //Configuration Group used by Sonar to provide Code Coverage using JaCoCo
 }
 
-// dependencies that are common across all java projects
 dependencies {
-	testCompile "cglib:cglib-nodep:$cglibVersion"
+	compile     "org.springframework.integration:spring-integration-core:$springIntegrationVersion"
+	testCompile "org.springframework.integration:spring-integration-test:$springIntegrationVersion"
 	testCompile "junit:junit-dep:$junitVersion"
 	testCompile "log4j:log4j:$log4jVersion"
-	testCompile "org.easymock:easymock:$easymockVersion"
-	testCompile "org.easymock:easymockclassextension:$easymockVersion"
-	testCompile "org.hamcrest:hamcrest-all:1.1"
 	testCompile "org.mockito:mockito-all:$mockitoVersion"
 	testCompile "org.springframework:spring-test:$springVersion"
 	jacoco group: "org.jacoco", name: "org.jacoco.agent", version: "0.5.6.201201232323", classifier: "runtime"
 }
+
 
 // enable all compiler warnings; individual projects may customize further
 ext.xLintArg = '-Xlint:all'
@@ -165,19 +132,16 @@ sonar {
 
 task api(type: Javadoc) {
 	group = 'Documentation'
-	description = 'Generates aggregated Javadoc API documentation.'
+	description = 'Generates the Javadoc API documentation.'
 	title = "${rootProject.description} ${version} API"
 	options.memberLevel = org.gradle.external.javadoc.JavadocMemberLevel.PROTECTED
 	options.author = true
 	options.header = rootProject.description
 	options.overview = 'src/api/overview.html'
-	source subprojects.collect { project ->
-		project.sourceSets.main.allJava
-	}
+
+	source = sourceSets.main.allJava
+	classpath = project.sourceSets.main.compileClasspath
 	destinationDir = new File(buildDir, "api")
-	classpath = files(subprojects.collect { project ->
-		project.sourceSets.main.compileClasspath
-	})
 }
 
 task schemaZip(type: Zip) {
@@ -186,46 +150,42 @@ task schemaZip(type: Zip) {
 	description = "Builds -${classifier} archive containing all " +
 		"XSDs for deployment at static.springframework.org/schema."
 
-	subprojects.each { subproject ->
-		def Properties schemas = new Properties();
-		def shortName = subproject.name.replaceFirst("${rootProject.name}-", '')
-		if (subproject.name.endsWith("-core")) {
-			shortName = ''
+	def Properties schemas = new Properties();
+	def shortName = idPrefix.replaceFirst("${idPrefix}-", '')
+
+	project.sourceSets.main.resources.find {
+		it.path.endsWith('META-INF/spring.schemas')
+	}?.withInputStream { schemas.load(it) }
+
+	for (def key : schemas.keySet()) {
+		File xsdFile = project.sourceSets.main.resources.find {
+			it.path.endsWith(schemas.get(key))
 		}
-
-		subproject.sourceSets.main.resources.find {
-			it.path.endsWith('META-INF/spring.schemas')
-		}?.withInputStream { schemas.load(it) }
-
-		for (def key : schemas.keySet()) {
-			File xsdFile = subproject.sourceSets.main.resources.find {
-				it.path.endsWith(schemas.get(key))
-			}
-			assert xsdFile != null
-			into ("integration/${shortName}") {
-				from xsdFile.path
-			}
+		assert xsdFile != null
+		into ("integration/${shortName}") {
+			from xsdFile.path
 		}
 	}
+
 }
 
 task docsZip(type: Zip) {
-		group = 'Distribution'
-		classifier = 'docs'
-		description = "Builds -${classifier} archive containing api and reference " +
+	group = 'Distribution'
+	classifier = 'docs'
+	description = "Builds -${classifier} archive containing api and reference " +
 			"for deployment at static.springframework.org/spring-integration/docs."
 
-		from('src/dist') {
-			include 'changelog.txt'
-		}
+	from('src/dist') {
+		include 'changelog.txt'
+	}
 
-		from (api) {
-			into 'api'
-		}
+	from (api) {
+		into 'api'
+	}
 
-		from (reference) {
-			into 'reference'
-		}
+	from (reference) {
+		into 'reference'
+	}
 }
 
 task distZip(type: Zip, dependsOn: [docsZip, schemaZip]) {
@@ -251,12 +211,10 @@ task distZip(type: Zip, dependsOn: [docsZip, schemaZip]) {
 		into "${baseDir}/schema"
 	}
 
-	subprojects.each { subproject ->
-		into ("${baseDir}/libs") {
-			from subproject.jar
-			from subproject.sourcesJar
-			from subproject.javadocJar
-		}
+	into ("${baseDir}/libs") {
+		from project.jar
+		from project.sourcesJar
+		from project.javadocJar
 	}
 }
 
@@ -272,14 +230,13 @@ task depsZip(type: Zip, dependsOn: distZip) { zipTask ->
 
 	gradle.taskGraph.whenReady { taskGraph ->
 		if (taskGraph.hasTask(":${zipTask.name}")) {
-			def projectNames = rootProject.subprojects*.name
+			def projectName = rootProject.name
 			def artifacts = new HashSet()
-			subprojects.each { subproject ->
-				subproject.configurations.runtime.resolvedConfiguration.resolvedArtifacts.each { artifact ->
-					def dependency = artifact.moduleVersion.id
-					if (!projectNames.contains(dependency.name)) {
-						artifacts << artifact.file
-					}
+
+			rootProject.configurations.runtime.resolvedConfiguration.resolvedArtifacts.each { artifact ->
+				def dependency = artifact.moduleVersion.id
+				if (!projectName.equals(dependency.name)) {
+					artifacts << artifact.file
 				}
 			}
 
@@ -303,5 +260,5 @@ task dist(dependsOn: assemble) {
 
 task wrapper(type: Wrapper) {
 	description = 'Generates gradlew[.bat] scripts'
-	gradleVersion = '1.0-rc-3'
+	gradleVersion = '1.1'
 }
