+++ si-template-projects/adapter/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,6 @@
-#Thu Jun 07 23:46:26 EDT 2012
+#Wed Sep 05 13:34:36 EDT 2012
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 zipStoreBase=GRADLE_USER_HOME
 zipStorePath=wrapper/dists
-distributionUrl=http\://services.gradle.org/distributions/gradle-1.0-rc-3-bin.zip
+distributionUrl=http\://services.gradle.org/distributions/gradle-1.1-bin.zip
