+++ si-template-projects/adapter/publish-maven.gradle
@@ -31,7 +31,7 @@ def customizePom(pom, gradleProject) {
 		generatedPom.project {
 			name = gradleProject.description
 			description = gradleProject.description
-			url = 'https://github.com/SpringSource/spring-integration'
+			url = 'https://github.com/SpringSource/spring-integration-extensions'
 			organization {
 				name = 'SpringSource'
 				url = 'http://springsource.org'
@@ -44,9 +44,9 @@ def customizePom(pom, gradleProject) {
 				}
 			}
 			scm {
-				url = 'https://github.com/SpringSource/spring-integration'
-				connection = 'scm:git:git://github.com/SpringSource/spring-integration'
-				developerConnection = 'scm:git:git://github.com/SpringSource/spring-integration'
+				url = 'https://github.com/SpringSource/spring-integration-extensions'
+				connection = 'scm:git:git://github.com/SpringSource/spring-integration-extensions'
+				developerConnection = 'scm:git:git://github.com/SpringSource/spring-integration-extensions'
 			}
 
 			developers {
