+++ si-template-projects/adapter/src/main/java/sipackage/config/xml/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides parser classes to provide Xml namespace support for the SIAdapterUpperPrefix components.
+ */
+package sipackage.config.xml;
