+++ si-template-projects/adapter/src/main/java/sipackage/core/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides core classes of the SIAdapterUpperPrefix module.
+ */
+package sipackage.core;
