+++ si-template-projects/adapter/src/main/java/sipackage/inbound/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides inbound Spring Integration SIAdapterUpperPrefix components.
+ */
+package sipackage.inbound;
