+++ si-template-projects/adapter/src/main/java/sipackage/outbound/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides Spring Integration components for doing outbound operations. 
+ */
+package sipackage.outbound;
