+++ si-template-projects/adapter/src/main/java/sipackage/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Root package of the SIAdapterUpperPrefix Module.
+ */
+package sipackage;
