+++ si-template-projects/adapter/src/main/java/sipackage/support/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides various support classes used across Spring Integration SIAdapterUpperPrefix Components.
+ */
+package sipackage.support;
