--- si-template-projects/adapter/src/main/java/stsorg/stsspring/stsint/config/xml/SIAdapterUpperPrefixNamespaceHandler.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2002-2012 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package stsorg.stsspring.stsint.config.xml;
-
-import org.springframework.integration.config.xml.AbstractIntegrationNamespaceHandler;
-
-/**
- * The namespace handler for the SIAdapterUpperPrefix namespace
- *
- * @author SI-TEMPLATE-AUTHOR
- * @since SI-TEMPLATE-VERSION
- *
- */
-public class SIAdapterUpperPrefixNamespaceHandler extends AbstractIntegrationNamespaceHandler {
-
-	/* (non-Javadoc)
-	 * @see org.springframework.beans.factory.xml.NamespaceHandler#init()
-	 */
-	public void init() {
-		this.registerBeanDefinitionParser("inbound-channel-adapter",  new SIAdapterUpperPrefixInboundChannelAdapterParser());
-		this.registerBeanDefinitionParser("outbound-channel-adapter", new SIAdapterUpperPrefixOutboundChannelAdapterParser());
-		this.registerBeanDefinitionParser("outbound-gateway", new SIAdapterUpperPrefixOutboundGatewayParser());
-	}
-}
