--- si-template-projects/adapter/src/main/java/stsorg/stsspring/stsint/config/xml/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Provides parser classes to provide Xml namespace support for the SIAdapterUpperPrefix components.
- */
-package stsorg.stsspring.stsint.config.xml;
