--- si-template-projects/adapter/src/main/java/stsorg/stsspring/stsint/core/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Provides core classes of the SIAdapterUpperPrefix module.
- */
-package stsorg.stsspring.stsint.core;
