--- si-template-projects/adapter/src/main/java/stsorg/stsspring/stsint/inbound/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Provides inbound Spring Integration SIAdapterUpperPrefix components.
- */
-package stsorg.stsspring.stsint.inbound;
