--- si-template-projects/adapter/src/main/java/stsorg/stsspring/stsint/outbound/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Provides Spring Integration components for doing outbound operations. 
- */
-package stsorg.stsspring.stsint.outbound;
