--- si-template-projects/adapter/src/main/java/stsorg/stsspring/stsint/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Root package of the SIAdapterUpperPrefix Module.
- */
-package stsorg.stsspring.stsint;
