--- si-template-projects/adapter/src/main/java/stsorg/stsspring/stsint/support/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Provides various support classes used across Spring Integration SIAdapterUpperPrefix Components.
- */
-package stsorg.stsspring.stsint.support;
