+++ si-template-projects/adapter/src/main/resources/META-INF/spring.schemas
@@ -1,2 +1,2 @@
-http\://www.springframework.org/schema/integration/siAdapterLowerPrefix/spring-integration-siAdapterLowerPrefix-2.2.xsd=org/springframework/integration/siAdapterLowerPrefix/config/xml/spring-integration-siAdapterLowerPrefix-2.2.xsd
-http\://www.springframework.org/schema/integration/siAdapterLowerPrefix/spring-integration-siAdapterLowerPrefix.xsd=stsorg/stsspring/stsint/config/xml/spring-integration-siAdapterLowerPrefix-2.2.xsd
+http\://www.springframework.org/schema/integration/siAdapterLowerPrefix/spring-integration-siAdapterLowerPrefix-1.0.xsd=org/springframework/integration/config/xml/spring-integration-siAdapterLowerPrefix-1.0.xsd
+http\://www.springframework.org/schema/integration/siAdapterLowerPrefix/spring-integration-siAdapterLowerPrefix.xsd=org/springframework/integration/config/xml/spring-integration-siAdapterLowerPrefix-1.0.xsd
