+++ si-template-projects/adapter/src/main/resources/org/springframework/integration/config/xml/spring-integration-siAdapterLowerPrefix.gif
