+++ si-template-projects/adapter/src/test/java/sipackage/config/xml/SIAdapterUpperPrefixMessageHandlerParserTests.xml
