+++ si-template-projects/adapter/src/test/java/sipackage/config/xml/SIAdapterUpperPrefixOutboundGatewayParserTests.xml
