+++ si-template-projects/adapter/src/test/resources/log4j.properties
@@ -5,4 +5,4 @@ log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
 log4j.appender.stdout.layout.ConversionPattern=%d{HH:mm:ss.SSS} %-5p [%t][%c] %m%n
 
 log4j.category.org.springframework.integration=WARN
-log4j.category.org.springframework.integration.jpa=INFO
+log4j.category.sipackage=INFO
