+++ si-template-projects/si-template-war-project/src/main/resources/rebel.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<application xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://www.zeroturnaround.com" xsi:schemaLocation="http://www.zeroturnaround.com http://www.zeroturnaround.com/alderaan/rebel-2_0.xsd">
+
+	<classpath>
+		<dir name="/Users/ghillert/dev/git-ghillert/spring-integration-templates/si-template-projects/si-template-war-project/target/classes">
+		</dir>
+		<dir name="/Users/ghillert/dev/git-ghillert/spring-integration-templates/si-template-projects/si-template-war-project/target/test-classes">
+		</dir>
+	</classpath>
+
+	<web>
+		<link target="/">
+			<dir name="/Users/ghillert/dev/git-ghillert/spring-integration-templates/si-template-projects/si-template-war-project/target/m2e-wtp/web-resources">
+			</dir>
+		</link>
+		<link target="/">
+			<dir name="/Users/ghillert/dev/git-ghillert/spring-integration-templates/si-template-projects/si-template-war-project/src/main/webapp">
+			</dir>
+		</link>
+	</web>
+
+</application>
